Fixing submodule debaucle once and for all.
If anyone is reading this, you are probably either laughing or swearing
at me. Or both.

I swear to god the idea of submodules clicked in my head fairly quickly,
but it's taken some time for me to actually get them to work correctly.
(As you can see.)

I really hope this works.